Move search() functionality to search_templates() in templates module
