+++ qubesadmin/templates.py
@@ -21,19 +21,20 @@
 
 import os
 import base64
+import collections
 import configparser
 import datetime
 import enum
 import fcntl
 import fnmatch
 import functools
+import itertools
 import logging
 import re
 import rpm
 import subprocess
 import tempfile
 import typing
-import sys
 
 import qubesadmin.app
 import qubesadmin.exc
@@ -49,6 +50,20 @@ PACKAGE_NAME_PREFIX = 'qubes-template-'
 TAR_HEADER_BYTES = 512
 WRAPPER_PAYLOAD_BEGIN = "###!Q!BEGIN-QUBES-WRAPPER!Q!###"
 WRAPPER_PAYLOAD_END = "###!Q!END-QUBES-WRAPPER!Q!###"
+
+# Search weight constants
+WEIGHT_NAME_EXACT = 1 << 4
+WEIGHT_NAME = 1 << 3
+WEIGHT_SUMMARY = 1 << 2
+WEIGHT_DESCRIPTION = 1 << 1
+WEIGHT_URL = 1 << 0
+
+WEIGHT_TO_FIELD = [
+    (WEIGHT_NAME_EXACT, 'Name'),
+    (WEIGHT_NAME, 'Name'),
+    (WEIGHT_SUMMARY, 'Summary'),
+    (WEIGHT_DESCRIPTION, 'Description'),
+    (WEIGHT_URL, 'URL')]
 
 
 class TemplateState(enum.Enum):
@@ -321,11 +336,9 @@ def extract_rpm(name: str, path: str, target: str) -> bool:
             return False
         link_path = f'{target}/{PATH_PREFIX}/{name}/template.rpm'
         try:
-            os.symlink(os.path.abspath(path),
-                       f'{target}/{PATH_PREFIX}/{name}/template.rpm')
+            os.symlink(os.path.abspath(path), link_path)
         except OSError as e:
-            print(f"Failed to create {link_path} symlink: {e!s}",
-                  file=sys.stderr)
+            log.error(f"Failed to create {link_path} symlink: {e!s}")
             return False
     return True
 
@@ -519,7 +532,7 @@ def qrexec_repoquery(
     proc.stderr.close()
     if proc.wait() != 0:
         for line in stderr.rstrip().split('\n'):
-            print(f"[Qrexec] {line}", file=sys.stderr)
+            log.error(f"[Qrexec] {line}")
         raise ConnectionError("qrexec call 'qubes.TemplateSearch' failed.")
     name_re = re.compile(r'\A[A-Za-z0-9._+][A-Za-z0-9._+-]*\Z')
     evr_re = re.compile(r'\A[A-Za-z0-9._+~]*\Z')
@@ -1032,3 +1045,81 @@ def migrate_from_rpmdb(app: qubesadmin.app.QubesBase):
     subprocess.check_call(
         ['rpm', '-e', '--justdb'] +
         [p[rpm.RPMTAG_NAME] for p in pkgs_to_remove])
+
+
+def search_templates(
+        app: qubesadmin.app.QubesBase,
+        repos: typing.List[typing.Tuple[str, str]],
+        releasever: str,
+        repo_files: typing.List[str],
+        keywords: typing.List[str],
+        updatevm: typing.Optional[str] = None,
+        search_all: bool = False,
+) -> typing.Tuple[
+        typing.List[Template],
+        typing.List[typing.Tuple[int, typing.List[typing.Tuple[int, str, bool]]]]]:
+    """Search template details for given patterns.
+
+    :param app: Qubes application object
+    :param repos: List of (repo_name, repo_url) tuples
+    :param releasever: Qubes release version for repo URLs
+    :param repo_files: List of paths to repo files
+    :param keywords: List of keywords to search for
+    :param updatevm: Name of the VM to use for updates (or None for default)
+    :param search_all: If True, search description and URL fields, and allow
+        partial keyword matches. If False, only search name and summary,
+        and require all keywords to match.
+    :return: Tuple of (query_res, search_res) where query_res is the list of
+        templates and search_res is a sorted list of (index, matches) tuples
+        where matches is a list of (weight, keyword, is_exact) tuples.
+    """
+    # Search in both installed and available templates
+    query_res = qrexec_repoquery(app, repos, releasever,
+                                 repo_files, updatevm)
+    for vm in app.domains:
+        if is_managed_template(vm):
+            query_res.append(query_local(vm))
+
+    # Get latest version for each template
+    query_res_tmp = []
+    for _, grp in itertools.groupby(sorted(query_res), lambda x: x[0]):
+        def compare(lhs, rhs):
+            return lhs if rpm.labelCompare(lhs[1:4], rhs[1:4]) > 0 else rhs
+
+        query_res_tmp.append(functools.reduce(compare, grp))
+    query_res = query_res_tmp
+
+    search_res_by_idx: \
+        typing.Dict[int, typing.List[typing.Tuple[int, str, bool]]] = \
+        collections.defaultdict(list)
+    for keyword in keywords:
+        for idx, entry in enumerate(query_res):
+            needle_types = \
+                [(entry.name, WEIGHT_NAME), (entry.summary, WEIGHT_SUMMARY)]
+            if search_all:
+                needle_types += [(entry.description, WEIGHT_DESCRIPTION),
+                                 (entry.url, WEIGHT_URL)]
+            for key, weight in needle_types:
+                if fnmatch.fnmatch(key, '*' + keyword + '*'):
+                    exact = keyword == key
+                    if exact and weight == WEIGHT_NAME:
+                        weight = WEIGHT_NAME_EXACT
+                    search_res_by_idx[idx].append((weight, keyword, exact))
+
+    if not search_all:
+        keywords_set = set(keywords)
+        idxs = list(search_res_by_idx.keys())
+        for idx in idxs:
+            if keywords_set != set(x[1] for x in search_res_by_idx[idx]):
+                del search_res_by_idx[idx]
+
+    def key_func(x):
+        # ORDER BY weight DESC, list_of_needles ASC, name ASC
+        idx, needles = x
+        weight = sum(t[0] for t in needles)
+        name = query_res[idx][0]
+        return -weight, needles, name
+
+    search_res = sorted(search_res_by_idx.items(), key=key_func)
+
+    return query_res, search_res
